requirements.txt
absl-py==0.12.0
astunparse==1.6.3
awscli==1.19.13
boto==2.49.0
boto3==1.17.13
botocore==1.20.13
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.3.9
coverage==5.4
Cython==0.29.22
docutils==0.15.2
flatbuffers==1.12
frozendict==1.2
gast==0.3.3
gevent==21.1.2
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery==3.3.6
google-cloud-core
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
google-cloud-bigquery-datatransfer
google-cloud-storage-transfer
googleapis-common-protos
graphviz==0.16
greenlet==1.0.0
httplib2==0.19.0
idna==2.10
importlib-metadata==4.0.1
jmespath==0.10.0
#Keras-Preprocessing==1.1.2
Markdown==3.3.4
mock==4.0.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
s3transfer==0.3.4
six==1.15.0
termcolor==1.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
wordnet==0.0.1b2
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.2.0
